Added support for Decorative Image
