+++ packages/terra-image/src/Image.jsx
@@ -1,27 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import objectFitImages from 'object-fit-images'; // Added polyfill for IE.
-import classNames from 'classnames';
-import classNamesBind from 'classnames/bind';
-import ThemeContext from 'terra-theme-context';
-import styles from './Image.module.scss';
-
-const cx = classNamesBind.bind(styles);
-
-const ImageVariant = {
-  DEFAULT: 'default',
-  ROUNDED: 'rounded',
-  CIRCLE: 'circle',
-  THUMBNAIL: 'thumbnail',
-};
-
-const FitTypes = {
-  COVER: 'cover',
-  SCALEDOWN: 'scale-down',
-  FILL: 'fill',
-  CONTAIN: 'contain',
-  NONE: 'none',
-};
+import { ImageVariant, FitTypes } from './_BaseImage';
+import InformativeImage from './_InformativeImage';
+import DecorativeImage from './_DecorativeImage';
 
 const propTypes = {
   /**
@@ -37,9 +18,11 @@ const propTypes = {
    */
   isFluid: PropTypes.bool,
   /**
+   * ![IMPORTANT](https://badgen.net/badge/UX/Accessibility/blue)
    * The text content that specifies an alternative text for an image.
+   * `alt` prop helps to provide meaningfull context for images and should be used for creating informative images. For decorative images `alt` prop can be ignored.
    */
-  alt: PropTypes.string.isRequired,
+  alt: PropTypes.string,
   /**
    * A React element which will be displayed during loading and in case of src load failure.
    */
@@ -66,127 +49,15 @@ const propTypes = {
   fit: PropTypes.oneOf(['cover', 'scale-down', 'fill', 'contain', 'none']),
 };
 
-/* eslint-disable react/default-props-match-prop-types */
-const defaultProps = {
-  variant: 'default',
-  isFluid: false,
-  alt: ' ',
-  fit: 'fill',
+const Image = (props) => {
+  if (props.alt !== undefined) {
+    return <InformativeImage {...props} />;
+  }
+
+  return <DecorativeImage {...props} />;
 };
-/* eslint-enable react/default-props-match-prop-types */
-
-class Image extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.ImageRef = React.createRef();
-
-    this.state = {
-      isLoading: true,
-      isError: false,
-      prevPropsSrc: props.src,
-    };
-
-    this.handleOnLoad = this.handleOnLoad.bind(this);
-    this.handleOnError = this.handleOnError.bind(this);
-  }
-
-  componentDidUpdate() {
-    objectFitImages(this.ImageRef.current);
-  }
-
-  static getDerivedStateFromProps(nextProps, prevState) {
-    if (nextProps.src !== prevState.prevPropsSrc) {
-      return {
-        isLoading: true,
-        isError: false,
-        prevPropsSrc: nextProps.src,
-      };
-    }
-
-    return null;
-  }
-
-  handleOnLoad() {
-    this.setState({ isLoading: false });
-    const { onLoad } = this.props;
-
-    if (onLoad !== undefined) {
-      onLoad();
-    }
-  }
-
-  handleOnError() {
-    this.setState({ isLoading: false, isError: true });
-    const { onError } = this.props;
-
-    if (onError !== undefined) {
-      onError();
-    }
-  }
-
-  createImage(customProps, imageClasses) {
-    const {
-      src, alt, height, width,
-    } = this.props;
-    return (
-      <img
-        {...customProps}
-        src={src}
-        alt={alt}
-        height={height}
-        width={width}
-        onLoad={this.handleOnLoad}
-        onError={this.handleOnError}
-        className={imageClasses}
-        ref={this.ImageRef}
-      />
-    );
-  }
-
-  render() {
-    const {
-      src, variant, isFluid, alt, placeholder, height, width, onLoad, onError, fit, ...customProps
-    } = this.props;
-
-    const theme = this.context;
-
-    const imageClasses = classNames(
-      cx(
-        'image',
-        fit,
-        variant,
-        theme.className,
-        { fluid: isFluid },
-        { hidden: placeholder && this.state.isLoading },
-      ),
-      customProps.className,
-    );
-
-    delete customProps.className;
-    if (!this.state.isLoading) {
-      objectFitImages(this.ImageRef.current);
-    }
-    if (placeholder) {
-      if (this.state.isLoading) {
-        return (
-          <>
-            {this.createImage(customProps, imageClasses)}
-            {placeholder}
-          </>
-        );
-      }
-
-      return this.state.isError ? placeholder : this.createImage(customProps, imageClasses);
-    }
-
-    return this.createImage(customProps, imageClasses);
-  }
-}
 
 Image.propTypes = propTypes;
-Image.defaultProps = defaultProps;
-Image.contextType = ThemeContext;
 
 export default Image;
 export { ImageVariant, FitTypes };
